postprocess: rewrite some subfixes checks

+++ lib/svtplay_dl/postprocess/__init__.py
@@ -4,13 +4,14 @@ import os
 import platform
 import re
 from requests import post, codes, Timeout
+from re import match
 
 from svtplay_dl.log import log
 from svtplay_dl.utils import which, run_program
 
 
 class postprocess(object):
-    def __init__(self, stream, options, subfixes=[]):
+    def __init__(self, stream, options, subfixes=None):
         self.stream = stream
         self.merge_subtitle = options.merge_subtitle
         self.external_subtitle = options.subtitle
@@ -68,12 +69,11 @@ class postprocess(object):
             'meankieli': 'fit',
             'jiddisch': 'yid'
         }
-        if len(self.subfixes) >= 2:
+        if self.subfixes and len(self.subfixes) >= 2:
             log.info("Determining the languages of the subtitles.")
         else:
             log.info("Determining the language of the subtitle.")
         if self.get_all_subtitles:
-            from re import match
             for subfix in self.subfixes:
                 if [exceptions[key] for key in exceptions.keys() if match(key, subfix.strip('-'))]:
                     if 'oversattning' in subfix.strip('-'):
@@ -140,7 +140,7 @@ class postprocess(object):
 
             if self.merge_subtitle and not self.external_subtitle:
                 log.info("Muxing done, removing the old files.")
-                if len(self.subfixes) >= 2:
+                if self.subfixes and len(self.subfixes) >= 2:
                     for subfix in self.subfixes:
                         subfile = "{0}.srt".format(name + subfix)
                         os.remove(subfile)
@@ -184,7 +184,7 @@ class postprocess(object):
             for stream_num, language in enumerate(langs, start=audiotrack + 1):
                 arguments += ["-map", "{}".format(videotrack), "-map", "{}".format(audiotrack), "-map", str(stream_num), "-c:s:" + str(stream_num - 2), "mov_text",
                               "-metadata:s:s:" + str(stream_num - 2), "language=" + language]
-            if len(self.subfixes) >= 2:
+            if self.subfixes and len(self.subfixes) >= 2:
                 for subfix in self.subfixes:
                     subfile = "{0}.srt".format(name + subfix)
                     cmd += ["-i", subfile]
@@ -202,7 +202,7 @@ class postprocess(object):
         os.remove(orig_filename)
         os.remove(audio_filename)
         if self.merge_subtitle and not self.external_subtitle:
-            if len(self.subfixes) >= 2:
+            if self.subfixes and len(self.subfixes) >= 2:
                 for subfix in self.subfixes:
                     subfile = "{0}.srt".format(name + subfix)
                     os.remove(subfile)
